+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/controller/CompanyController.java
@@ -4,8 +4,8 @@ import kr.ac.kumoh.d138.JobForeigner.global.response.GlobalPageResponse;
 import kr.ac.kumoh.d138.JobForeigner.global.response.ResponseBody;
 import kr.ac.kumoh.d138.JobForeigner.global.response.ResponseUtil;
 import kr.ac.kumoh.d138.JobForeigner.job.api.CompanyApi;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.CompanyDetailResponseDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.CompanyResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.CompanyDetailResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.CompanyResponseDto;
 import kr.ac.kumoh.d138.JobForeigner.job.service.CompanyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
