+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/controller/JobPostController.java
@@ -4,11 +4,11 @@ import jakarta.validation.Valid;
 import kr.ac.kumoh.d138.JobForeigner.global.response.GlobalPageResponse;
 import kr.ac.kumoh.d138.JobForeigner.global.response.ResponseBody;
 import kr.ac.kumoh.d138.JobForeigner.global.response.ResponseUtil;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.request.JobPostRequestDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.request.JobTempPostRequestDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.response.JobPostDetailResponseDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.response.JobPostResponseDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.response.UpdateJobPostResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.request.JobPostRequestDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.request.JobTempPostRequestDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.JobPostDetailResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.JobPostResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.UpdateJobPostResponseDto;
 import kr.ac.kumoh.d138.JobForeigner.job.service.JobPostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
