+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/service/CompanyService.java
@@ -8,7 +8,13 @@ import kr.ac.kumoh.d138.JobForeigner.job.domain.Company;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.CompanyRating;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.JobPost;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.QCompany;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.*;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.CompanyDetailResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.CompanyInfoDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.CompanyRatingDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.CompanyResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.JobPostDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.ReviewDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.SalaryInfoDto;
 import kr.ac.kumoh.d138.JobForeigner.job.repository.CompanyRepository;
 import kr.ac.kumoh.d138.JobForeigner.job.repository.JobPostRepository;
 import kr.ac.kumoh.d138.JobForeigner.rating.Rating;
