+++ src/main/java/kr/ac/kumoh/d138/JobForeigner/job/service/JobPostService.java
@@ -6,11 +6,11 @@ import jakarta.transaction.Transactional;
 import kr.ac.kumoh.d138.JobForeigner.global.exception.BusinessException;
 import kr.ac.kumoh.d138.JobForeigner.global.exception.ExceptionType;
 import kr.ac.kumoh.d138.JobForeigner.job.domain.*;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.request.JobPostRequestDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.request.JobTempPostRequestDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.response.JobPostDetailResponseDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.response.JobPostResponseDto;
-import kr.ac.kumoh.d138.JobForeigner.job.dto.company.response.UpdateJobPostResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.request.JobPostRequestDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.request.JobTempPostRequestDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.JobPostDetailResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.JobPostResponseDto;
+import kr.ac.kumoh.d138.JobForeigner.job.dto.response.UpdateJobPostResponseDto;
 import kr.ac.kumoh.d138.JobForeigner.job.repository.CompanyRepository;
 import kr.ac.kumoh.d138.JobForeigner.job.repository.JobApplicationRepository;
 import kr.ac.kumoh.d138.JobForeigner.job.repository.JobPostRepository;
